TemplateMatching

Documentation for TemplateMatching.

TemplateMatching.correlatetemplate

TemplateMatching.crosscorrelate

TemplateMatching.maxfilter

TemplateMatching.stack

TemplateMatching.correlatetemplate — Function

correlatetemplate(data, template, shifts, tolerance, [cc_eltype=Float64])

Return the cross-correlation between data and template. The cross-correlations for each series are aligned by shifts and averaged. If toleranceis not zero, then the average accounts for possible misplacement of each series bytolerance` sample, and return the average of the maximum cross-correlation compatible with that misplacement.

TemplateMatching.crosscorrelate — Method

crosscorrelate(series, template, [cc_eltype=Float64], normalize_template=true)

Return the normalized cross-correlation between series and template.

If series is a vector of $n$ elements and template a vector of $N$ elements, cross-correlation will be a vector of $n - N + 1$ elements of type cc_eltype. The element of cross-correlation $χ_k$ is Pearson correlation coefficient between $x^{(k)}_i$ denoting the elements of series $x_{i + k}$ and the elements of template $y_i$, with $i = 1, 2, \dots, N$

$χ_k = \frac{\sum{\left(x^{(k)}_i - μ_{x^{(k)}}\right) \left(y_i - μ_y\right)}}{N σ_{x^{(k)}} σ_y} \, ,$

where $μ$ and $σ$ denotes the mean and variance respectively.

If normalize_template is set to false then template is assumed to have mean and std respectively equal to 0.0 and 1.0.

See the Wikipedia page on cross-correlation for more details.

Examples

julia> crosscorrelate(sin.(0:0.25pi:2pi), [1, 1+√2, 1])
7-element Vector{Float64}:
  0.23258781949447394
  1.000000000000001
  0.23258781949447402
  7.401486830834377e-17
 -0.23258781949447394
 -1.0000000000000007
 -0.23258781949447394

TemplateMatching.maxfilter — Method

maxfilter(x, tolerance)

Return a vector of the same size of x whose element at index n is the maximum of the elements of x which are at most tolerance apart from x[n] (i.e. in a window of at most 2 * tolerance + 1 elements).

Examples

julia> maxfilter(sin.(0:0.25pi:2pi), 1)
9-element Vector{Float64}:
  0.7071067811865475
  1.0
  1.0
  1.0
  0.7071067811865476
  1.2246467991473532e-16
 -0.7071067811865475
 -2.4492935982947064e-16
 -2.4492935982947064e-16

TemplateMatching.stack — Method

stack(correlations, offsets)

Return the average cross-correlation after aligning correlations. Each cross-correlation correlations[n] is shifted to the right by offsets[n].

Examples

julia> stack([[0, 1.0, 0, 0], [0, 0, 1.0, 0]], [1, 2])
3-element OffsetArray(::Vector{Float64}, 0:2) with eltype Float64 with indices 0:2:
 0.0
 1.0
 0.0