TemplateMatching

Documentation for TemplateMatching.

TemplateMatching.correlatetemplate

TemplateMatching.crosscorrelate

TemplateMatching.findmax_window

TemplateMatching.findpeaks

TemplateMatching.mad_test

TemplateMatching.magnitude

TemplateMatching.maxfilter

TemplateMatching.stack

TemplateMatching.correlatetemplate — Function

correlatetemplate(data, template, offsets, tolerance, [element_type=Float64])

Return the cross-correlation between data and template. The cross-correlations for each series are aligned using offsets and averaged. If toleranceis not zero, then the average accounts for possible misplacement of each series bytolerance` sample, and return the average of the maximum cross-correlation compatible with that misplacement.

TemplateMatching.crosscorrelate — Method

crosscorrelate(series, template, [element_type=Float64], normalize_template=true)

Return the normalized cross-correlation between series and template.

If series is a vector of $n$ elements and template a vector of $N$ elements, cross-correlation will be a vector of $n - N + 1$ elements of type element_type. The element of cross-correlation $χ_k$ is Pearson correlation coefficient between $x^{(k)}_i$ denoting the elements of series $x_{i + k}$ and the elements of template $y_i$, with $i = 1, 2, \dots, N$

$χ_k = \frac{\sum{\left(x^{(k)}_i - μ_{x^{(k)}}\right) \left(y_i - μ_y\right)}}{N σ_{x^{(k)}} σ_y} \, ,$

where $μ$ and $σ$ denotes the mean and variance respectively.

If normalize_template is set to false then template is assumed to have mean and std respectively equal to 0.0 and 1.0.

See the Wikipedia page on cross-correlation for more details.

Examples

julia> crosscorrelate(sin.(0:0.25pi:2pi), [1, 1+√2, 1])
7-element Vector{Float64}:
  0.23258781949447394
  1.000000000000001
  0.23258781949447402
  7.401486830834377e-17
 -0.23258781949447394
 -1.0000000000000007
 -0.23258781949447394

TemplateMatching.findmax_window — Method

findmax_window(crosscorrelation, index, tolerance)

Return a tuples containing the maximum of crosscorrelation, in a window centered around index and radius equal to tolerace, and its index.

Examples

julia> findmax_window.([abs.(sin.(0:0.01pi:4pi)) for _ = 1:4], [49, 153, 249, 353], 5)
4-element Vector{Tuple{Float64, Int64}}:
 (1.0, 51)
 (1.0, 151)
 (1.0, 251)
 (1.0, 351)

TemplateMatching.findpeaks — Method

findpeaks(signal, threshold, distance)

Find all the local maxima of signal above threshold and at least distance apart. The maxima closer than distance are removed starting from the ones with higher neighbour maxima up until the condition is met.

Return two vectors one with indices and one with heights of the peaks.

The implementation is taken from scipy.signal.

Examples

julia> findpeaks(abs.(sin.(0:0.01pi:2pi)), 0.5, 0)
([51, 151], [1.0, 1.0])

TemplateMatching.mad_test — Function

mad_test(xs, [r=3])

Return a vector of booleans whose elements are true if the absolute deviation from the median of xs of the corresponding elements (of xs) is at least r times the median absolute deviation, and false otherwise.

Examples

julia> TemplateMatching.mad_test([1, 2, 3, 100])
4-element BitVector:
 0
 0
 0
 1

TemplateMatching.magnitude — Function

magnitude(data, template, indices, [outlier=TemplateMatching.mad], [avg=StatsBase.mean])

Return the average relative magnitude of the view of series in data starting at indices and having the same length of the corresponding template series. The outliers, i.e. the elements corresponding to those of outlier(magnitudes) evaluates that are true, are excluded. The average is calculated using avg(magnitudes).

TemplateMatching.maxfilter — Method

maxfilter(x, tolerance)

Return a vector of the same size of x whose element at index n is the maximum of the elements of x which are at most tolerance apart from x[n] (i.e. in a window of at most 2 * tolerance + 1 elements).

Examples

julia> maxfilter(sin.(0:0.25pi:2pi), 1)
9-element Vector{Float64}:
  0.7071067811865475
  1.0
  1.0
  1.0
  0.7071067811865476
  1.2246467991473532e-16
 -0.7071067811865475
 -2.4492935982947064e-16
 -2.4492935982947064e-16

TemplateMatching.stack — Method

stack(correlations, offsets)

Return the average cross-correlation after aligning correlations. Each cross-correlation correlations[n] is shifted to the right by offsets[n].

Examples

julia> stack([[0, 1.0, 0, 0], [0, 0, 1.0, 0]], [1, 2])
3-element OffsetArray(::Vector{Float64}, 0:2) with eltype Float64 with indices 0:2:
 0.0
 1.0
 0.0